Move InputCell to core.py
