+++ roxene/__init__.py
@@ -2,9 +2,7 @@ import tensorflow as tf
 
 PRECISION = tf.dtypes.float16
 
-from .core import Gene, Organism, Cell
+from .core import Gene, Organism, Cell, InputCell
 from .Neuron import Neuron
-from .InputCell import InputCell
-from .ConnectNeurons import ConnectNeurons
 from .ConnectNeurons import ConnectNeurons
 from .CreateNeuron import CreateNeuron
