+++ roxene/core.py
@@ -26,3 +26,12 @@ class Gene(abc.ABC):
     @abc.abstractmethod
     def execute(self, organism: Organism):
         pass
+
+
+class InputCell(Cell):
+
+    def __init__(self, initial_value):
+        self.value = initial_value
+
+    def get_output(self):
+        return self.value
